Network.TypedProtocol.PingPong.Server

Documentation

data PingPongServer m a

Constructors

PingPongServer

Fields

recvMsgPing ∷ m (PingPongServer m a)

The client sent us a ping message. We have no choices here, and the response is nullary, all we have are local effects.

recvMsgDone ∷ a

The client terminated. Here we have a pure return value, but we could have done another action in m if we wanted to.

pingPongServerPeer ∷ Monad m ⇒ PingPongServer m a → Peer PingPong AsServer StIdle m a

Interpret a particular server action sequence into the server side of the PingPong protocol.